ill continue this when im home
