+++ package.json
@@ -7,6 +7,7 @@
   },
   "devDependencies": {
     "@types/ms": "^0.7.32",
+    "@types/node": "^20.10.1",
     "bun-types": "latest"
   },
   "peerDependencies": {
