+++ src/Events/index.ts
@@ -1,0 +1,6 @@
+import GuildCreateEvent from "./guildCreate";
+import InteractionCreateEvent from "./interactionCreate";
+import MessageCreateEvent from "./messageCreate";
+import ReadyEvent from "./ready";
+
+export { GuildCreateEvent, InteractionCreateEvent, MessageCreateEvent, ReadyEvent };
