+++ src/Helpers/map.ts
@@ -1,7 +1,7 @@
 import { Client, EmbedBuilder, Message } from "discord.js";
 import { v2 } from "osu-api-extended";
 import { getBeatmap } from "../functions";
-import { BeatmapInfo } from "../types";
+import { BeatmapInfo } from "../Structure";
 import { downloadMap, getBeatmapId_FromContext, getMap, getUsernameFromArgs, insertData, Interactionhandler } from "../utils";
 
 export async function start({ interaction, client, args, mapId }: { interaction: Message; client?: Client<boolean>; args: string[]; mapId?: string }) {
