+++ src/Helpers/plays.ts
@@ -3,7 +3,7 @@ import { v2 } from "osu-api-extended";
 import { response as ScoreResponse } from "osu-api-extended/dist/types/v2_scores_user_category";
 import { MyClient } from "../classes";
 import { getScore, getUser } from "../functions";
-import { commands, osuModes, UserInfo } from "../types";
+import { commands, osuModes, UserInfo } from "../Structure";
 import { buildActionRow, buttonBoolsIndex, buttonBoolsTops, downloadMap, firstButton, getMap, getUsernameFromArgs, insertData, Interactionhandler, lastButton, nextButton, previousButton, specifyButton } from "../utils";
 
 export async function start({ isTops, interaction, passOnly: passOnlyArg, args, mode: modeArg, number, recentTop, client }: { isTops: boolean; interaction: Message | ChatInputCommandInteraction; passOnly?: boolean; args?: string[]; mode?: osuModes; number?: number; recentTop?: boolean; client: MyClient }) {
