+++ src/Structure/index.ts
@@ -1,0 +1,4 @@
+import ExtendedClient from "./ExtendedClient";
+
+export * from "./types";
+export { ExtendedClient };
